Moved login helpers out of test directory to general E2E testing.
