+++ src/js/hca/tests/hca-helpers.js
@@ -1,6 +1,6 @@
 const mock = require('../../../../test/e2e/mock-helpers');
 const Timeouts = require('../../../platform/testing/e2e/timeouts.js');
-const LoginHelpers = require('../../../platform/site-wide/login/tests/login-helpers.js');
+const LoginHelpers = require('../../../platform/testing/e2e/auth.js');
 const moment = require('moment');
 
 function completePersonalInformation(client, data) {
