+++ src/js/letters/tests/01-authed.e2e.spec.js
@@ -3,7 +3,7 @@ import startCase from 'lodash/startCase';
 const E2eHelpers = require('../../../platform/testing/e2e/helpers');
 const Timeouts = require('../../../platform/testing/e2e/timeouts.js');
 const LettersHelpers = require('./letters-helpers.js');
-const LoginHelpers = require('../../../platform/site-wide/login/tests/login-helpers');
+const LoginHelpers = require('../../../platform/testing/e2e/auth');
 
 const newAddress = LettersHelpers.newAddress.data.attributes.address;
 const oldAddress = LettersHelpers.address.data.attributes.address;
