+++ src/js/messaging/tests/01-required-desktop.e2e.spec.js
@@ -1,7 +1,7 @@
 const E2eHelpers = require('../../../platform/testing/e2e/helpers');
 const Timeouts = require('../../../platform/testing/e2e/timeouts.js');
 const MessagingHelpers = require('./messaging-helpers');
-const LoginHelpers = require('../../../platform/site-wide/login/tests/login-helpers');
+const LoginHelpers = require('../../../platform/testing/e2e/auth');
 const AccountCreationHelpers = require('../../../../test/e2e/account-creation-helpers');
 
 module.exports = E2eHelpers.createE2eTest(
