+++ src/js/post-911-gib-status/tests/01-authed.e2e.spec.js
@@ -1,6 +1,6 @@
 const E2eHelpers = require('../../../platform/testing/e2e/helpers');
 const Timeouts = require('../../../platform/testing/e2e/timeouts');
-const LoginHelpers = require('../../../platform/site-wide/login/tests/login-helpers');
+const LoginHelpers = require('../../../platform/testing/e2e/auth');
 const GibsHelpers = require('./post-911-gib-status-helpers');
 
 module.exports = E2eHelpers.createE2eTest(
