+++ src/js/user-profile/tests/e2e/user-profile.e2e.spec.js
@@ -1,6 +1,6 @@
 const E2eHelpers = require('../../../../platform/testing/e2e/helpers');
 const Timeouts = require('../../../../platform/testing/e2e/timeouts');
-const LoginHelpers = require('../../../../platform/site-wide/login/tests/login-helpers');
+const LoginHelpers = require('../../../../platform/testing/e2e/auth');
 
 const { formLinks } = require('../../helpers');
